==> analysis/1 Exploration.Rmd <==
---
title: "exploration"
author: "Jordan Boswell and Thuy Nguyen"
date: "2/26/2022"
output: html_document
---


```{r setup, include=FALSE}
# Thuy: '~/Documents/GitHub/spaceship-titanic'
# Jordan: 'C:/Users/jbos1/Desktop/Projects/Kaggle/spaceship-titanic'
# Jordan Laptop: 'C:/Users/User/Documents/Projects/Kaggle/spaceship-titanic'
library(visdat)
library(naniar)
library(nullabor)
library(ggplot2)
theme_set(theme_classic())
source('0 Functions.R')
knitr::opts_knit$set(echo = TRUE, root.dir = 'C:/Users/User/Documents/Projects/Kaggle/spaceship-titanic')
```

# Read data and format columns
```{r}
train <- read.csv("data/train.csv", na.strings=c("NA", ""))
test <- read.csv("data/test.csv", na.strings=c("NA", ""))
original_cols <- names(train)
train$Train <- TRUE
test$Train <- FALSE
test$Transported <- NA
ship <- rbind(train, test)
create_base_columns <- function(df){
    df$GID <- as.integer(substr(df$PassengerId, 1,4))
    df$IID <- as.integer(substr(df$PassengerId,6,8))
    df$HomePlanet <- as.factor(df$HomePlanet)
    df$CryoSleep <- as.logical(df$CryoSleep)
    df$Destination <- as.factor(df$Destination)
    df$Age <- as.integer(df$Age)
    df$VIP <- as.logical(df$VIP)
    df
}
ship <- create_base_columns(ship)
ship$Transported <- as.logical(ship$Transported)
```

# High-Level Visualization & Missing Data Exploration
```{r}
vis_dat(ship[, original_cols])
vis_miss(ship[, setdiff(original_cols, 'Transported')])
mcar_test(ship)
mcar_test(ship[, setdiff(original_cols, 'Transported')])
```

# Add Columns
## Deck/Side/Num
```{r}
ship <- updateDeckSideNum(ship)
```

## Spending
```{r}
ship <- updateSpending(ship)
```

## Last Name
```{r}
ship$LastName <- sapply(ship$Name, function(x){strsplit(x,split = " ")}[[1]][2])
```

# EDA
## GID
### GID values form a contiguous (integer) range from 1 to 9280
```{r}
for(i in min(ship$GID):max(ship$GID)){
  if (!(i %in% ship$GID)){
    print(i)
  }
}
```

### test$GID and train$GID are disjoint
```{r}
intersect(ship$GID[ship$Train], ship$GID[!ship$Train])
```

## CryoSleep implies 0 spending
### Evidence
```{r}
ftable(HasSpent~CryoSleep, data=ship)
```

### Update
```{r}
ship <- updateSpendingOfCryo(ship)
```

## Young children (Age <= 12) don't spend money
### Evidence
```{r}
spending_mean <- sapply(split(ship$Spending, ship$Age), mean)
spending_25 <- sapply(split(ship$Spending, ship$Age), function(x){quantile(x, 0.25)})
spending_75 <- sapply(split(ship$Spending, ship$Age), function(x){quantile(x, 0.75)})
spending_length <- sapply(split(ship$Spending, ship$Age), length)

plot(names(spending_mean), spending_mean, main='Spending by Age', xlab='Age', ylab='Spending', col='red', type='l', lwd=2, ylim=c(0, max(spending_75)), xaxp=c(0, 80, 8))
lines(names(spending_mean), spending_25, col='light blue', lwd=2)
lines(names(spending_mean), spending_75, col='dark blue', lwd=2)
barplot(spending_length*15, add=T, space=0, col=rgb(.5,.5,.5,0.07), border=rgb(.5,.5,.5,0.22), names.arg=NA)
abline(v=12, lty=2)
legend(x='topleft', legend=c('75th percentile', 'mean', '25th percentile'), col=c('dark blue', 'red', 'light blue'), lwd=2)

table(as.factor(ifelse(ship$Age <= 12, 1, 0)), ship$HasSpent)
```
### Update
```{r}
ship <- updateSpendingOfChildren(ship)
```

## Link Between Homeplanet and Deck: 
* Deck G implies Earth
* Deck A,B,C,T implies Europa
### Evidence
```{r}
ship$Deck <- factor(ship$Deck)
ship$Side <- factor(ship$Side)
table(ship$HomePlanet, ship$Deck)
```

### Update
```{r}
ship <- updateHomePlanetFromDeck(ship)
```

## Same group implies same side
### Evidence
```{r}
#check
uni_GID <- unique(ship$GID)
unique(sapply(uni_GID, function(i){length(unique(ship$Side[!is.na(ship$Side) & ship$GID == i]))}))
```

### Update
```{r}
ship <- updateSideFromGID(ship)
```

## Surnames are Unique to HomePlanet
### Evidence
```{r}
intersect(ship$LastName[!is.na(ship$LastName) & !is.na(ship$HomePlanet) &ship$HomePlanet =="Earth"], ship$LastName[!is.na(ship$LastName) & !is.na(ship$HomePlanet) & ship$HomePlanet == "Europa"])
intersect(ship$LastName[!is.na(ship$LastName) & !is.na(ship$HomePlanet) &ship$HomePlanet =="Earth"], ship$LastName[!is.na(ship$LastName) & !is.na(ship$HomePlanet) & ship$HomePlanet == "March"])
intersect(ship$LastName[!is.na(ship$LastName) & !is.na(ship$HomePlanet) &ship$HomePlanet =="March"], ship$LastName[!is.na(ship$LastName) & !is.na(ship$HomePlanet) & ship$HomePlanet == "Europa"])
```

### Update
```{r}
ship <- updateHomePlanetFromSurname(ship)
```

## Earth Doesn't Have VIP
### Evidence
```{r}
table(ship$HomePlanet, ship$VIP)
```

### Update
```{r}
ship <- updateVIPFromHomePlanet(ship)
```

## Deck
### Does CryoSleep affect Deck?
#### Distribution of CryoSleep
```{r}
signif(table(ship$Deck[!is.na(ship$CryoSleep) & !is.na(ship$Deck) & ship$CryoSleep == T]) / sum(!is.na(ship$CryoSleep) & !is.na(ship$Deck) & ship$CryoSleep == T) * 100, 4)
```
#### Per-deck percentage of CryoSleep
```{r}
signif(table(ship$Deck[!is.na(ship$CryoSleep) & !is.na(ship$Deck) & ship$CryoSleep == T]) / tapply(ship$Deck, ship$Deck, FUN=length) * 100, 4)
```

### Exploring Num within each Deck
```{r}
# todo - Change to multi-dist plot (e.g. violin)
for (deck in sort(unique(ship$Deck))){
    if (is.na(deck)) next
    hist(ship$Num[!is.na(ship$Deck) & ship$Deck == deck], main=deck)
}
```
### Room Size
```{r}
# Average number of people per room (for each deck)
tapply(ship$Num, ship$Deck, function(x){length(x)/length(unique(x))})

tapply(ship$Age, ship$Deck, summary)
```

## Group
### Create Separate Group Dataset
```{r}
updateGroups <- function(df) {
  splits <- split(df, df$GID)
  df_groups <- data.frame(t(sapply(splits, function(x) {
    c(
      GID = x$GID[1],
      GroupTrain = x$Train[1],
      GroupSize = nrow(x),
      GroupNumCabinsNA = nrow(subset(x, is.na(Cabin))),
      GroupNumCabins = ifelse(nrow(x) == 1, 1, length(unique(x$Cabin[!is.na(x$Cabin)]))),
      GroupNumTransported = sum(x$Transported, na.rm = T),
      GroupNumNotTransported = sum(x$Transported == FALSE, na.rm = T),
      #GroupTransportedPct = ifelse(sum(!is.na(x$Transported) == 0), NA, sum(x$Transported == T, na.rm=T)/sum(!is.na(x$Transported))),
      GroupCabinMode = ifelse(sum(is.na(x$Cabin)) == nrow(x), NA, names(sort(
        table(x$Cabin), decreasing = T
      ))[1]),
      GroupNumInCabinMode = ifelse(sum(is.na(x$Cabin)) == nrow(x), NA, sum(
        !is.na(x$Cabin) &
          x$Cabin == names(sort(table(x$Cabin), decreasing = T))[1]
      )),
      GroupNumHomePlanets = ifelse(nrow(x) == 1, 1, length(unique(x$HomePlanet[!is.na(x$HomePlanet)]))),
      GroupNumDestinations = ifelse(nrow(x) == 1, 1, length(unique(x$Destination[!is.na(x$Destination)]))),
      GroupHomePlanetMode = ifelse(sum(is.na(x$HomePlanet)) == nrow(x), NA, names(sort(
        table(x$HomePlanet), decreasing = T
      ))[1]),
      GroupDestinationMode = ifelse(sum(is.na(x$Destination)) == nrow(x), NA, names(sort(
        table(x$Destination), decreasing = T
      ))[1])
    )
  })))
  df_groups$GID <- as.integer(df_groups$GID)
  df_groups$GroupTrain <- as.logical(df_groups$GroupTrain)
  df_groups$GroupSize <- as.integer(df_groups$GroupSize)
  df_groups$GroupNumCabinsNA <- as.integer(df_groups$GroupNumCabinsNA)
  df_groups$GroupNumCabins <- as.integer(df_groups$GroupNumCabins)
  df_groups$GroupNumTransported <-
    as.integer(df_groups$GroupNumTransported)
  df_groups$GroupNumNotTransported <-
    as.integer(df_groups$GroupNumNotTransported)
  df_groups$GroupNumInCabinMode <-
    as.integer(df_groups$GroupNumInCabinMode)
  df_groups$GroupNumHomePlanets <-
    as.integer(df_groups$GroupNumHomePlanets)
  df_groups$GroupNumDestinations <-
    as.integer(df_groups$GroupNumDestinations)
  df <- merge(df, df_groups, by = 'GID', all.x = T)
  # todo
  return(list(df = df, df_groups = df_groups))
}
```

### Add Group Columns
```{r}
group_results <- updateGroups(ship)
ship <- group_results$df
ship_groups <- group_results$df_groups
```

## Cabin
### Impute Cabin Based on Group Cabin Mode
#### Evidence
*If a person has a group that is all in the same cabin, then should we expect that person to also be in that same cabin?*
```{r}
for (i in 2:max(ship_groups$GroupSize)){
    print(paste0("All Non NA Same Cabin (For Groups With Non-NA Size ", i, "):"))
    s <- ship_groups[ship_groups$GroupSize - ship_groups$GroupNumCabinsNA == i & (ship_groups$GroupNumCabins == 1 | (ship_groups$GroupNumCabins == 2 & ship_groups$GroupSize - (ship_groups$GroupNumInCabinMode + ship_groups$GroupNumCabinsNA) == 1)),]
    st <- table(s$GroupNumInCabinMode == s$GroupSize - s$GroupNumCabinsNA)
    print(st)
    num_true <- ifelse('TRUE' %in% names(st), st[['TRUE']], 0)
    num_false <- ifelse('FALSE' %in% names(st), st[['FALSE']], 0)
    p_same <- i * num_true / (i*num_true + num_false)
    print(paste0('Probability of same cabin: ', round(p_same,2)))
    print("-------------------------------------------")
}
rm(s)
```

#### Update
```{r}
ship$Cabin[is.na(ship$Cabin) & !is.na(ship$GroupCabinMode)] <- ship$GroupCabinMode[is.na(ship$Cabin) & !is.na(ship$GroupCabinMode)]
ship <- updateDeckSideNum(ship)
```

### Create Separate Cabin Dataset
```{r}
splits <- split(ship[!is.na(ship$Cabin), ], ship$Cabin[!is.na(ship$Cabin)])
ship_cabins <- data.frame(t(
    sapply(splits, function(x){c(
        CabinSize = length(x$PassengerId),
        Cabin = x$Cabin[1],
        Deck = x$Deck[1],
        Side = x$Side[1],
        Num = x$Num[1],
        CabinNumNATransported = sum(is.na(x$Transported)),
        CabinNumTransported = sum(x$Transported, na.rm=T),
        CabinNumNotTransported = sum(x$Transported==FALSE, na.rm=T),
        CabinTransportedPct = ifelse(sum(!is.na(x$Transported) == 0), NA, sum(x$Transported == T, na.rm=T)/sum(!is.na(x$Transported))),
        CabinNumCryo = sum(x$CryoSleep, na.rm=T)
    )})
))
ship_cabins$CabinSize <- as.integer(ship_cabins$CabinSize)
ship_cabins$Num <- as.integer(ship_cabins$Num)
ship_cabins$CabinNumNATransported <- as.integer(ship_cabins$CabinNumNATransported)
ship_cabins$CabinNumTransported <- as.integer(ship_cabins$CabinNumTransported)
ship_cabins$CabinNumNotTransported <- as.integer(ship_cabins$CabinNumNotTransported)
ship_cabins$CabinTransportedPct <- as.numeric(ship_cabins$CabinTransportedPct)
ship_cabins$CabinNumCryo <- as.integer(ship_cabins$CabinNumCryo)
ship_cabins$SideNeighbors <- as.integer(NA)
ship_cabins$SideNeighborsTransported <- as.integer(NA)
ship_cabins$SideNeighborsTransportedPct <- as.numeric(NA)
ship_cabins$FrontNeighbors <- as.integer(NA)
ship_cabins$FrontNeighborsTransported <- as.integer(NA)
ship_cabins$FrontNeighborsTransportedPct <- as.numeric(NA)
ship_cabins$BackNeighbors <- as.integer(NA)
ship_cabins$BackNeighborsTransported <- as.integer(NA)
ship_cabins$BackNeighborsTransportedPct <- as.numeric(NA)
ship_cabins$DiagFrontNeighbors <- as.integer(NA)
ship_cabins$DiagFrontNeighborsTransported <- as.integer(NA)
ship_cabins$DiagFrontNeighborsTransportedPct <- as.numeric(NA)
ship_cabins$DiagBackNeighbors <- as.integer(NA)
ship_cabins$DiagBackNeighborsTransported <- as.integer(NA)
ship_cabins$DiagBackNeighborsTransportedPct <- as.numeric(NA)
for (i in 1:nrow(ship_cabins)){
    side_neighbor <- subset(ship_cabins, Deck == Deck[i] & Num == Num[i] & Side != Side[i])
    back_neighbor <- subset(ship_cabins, Deck == Deck[i] & Num == Num[i]-1 & Side == Side[i])
    front_neighbor <- subset(ship_cabins, Deck == Deck[i] & Num == Num[i]+1 & Side == Side[i])
    diag_back_neighbor <- subset(ship_cabins, Deck == Deck[i] & Num == Num[i]-1 & Side != Side[i])
    diag_front_neighbor <- subset(ship_cabins, Deck == Deck[i] & Num == Num[i]+1 & Side != Side[i])
    ship_cabins$SideNeighbors[i] <- ifelse(nrow(side_neighbor) == 0, 0, side_neighbor$CabinNumTransported + side_neighbor$CabinNumNotTransported)
    ship_cabins$SideNeighborsTransported[i] <- ifelse(nrow(side_neighbor) == 0, 0, side_neighbor$CabinNumTransported)
    ship_cabins$SideNeighborsTransportedPct[i] <- ifelse(nrow(side_neighbor) == 0, 0, side_neighbor$CabinTransportedPct)
    ship_cabins$FrontNeighbors[i] <- ifelse(nrow(front_neighbor) == 0, 0, front_neighbor$CabinNumTransported + front_neighbor$CabinNumNotTransported)
    ship_cabins$FrontNeighborsTransported[i] <- ifelse(nrow(front_neighbor) == 0, 0, front_neighbor$CabinNumTransported)
    ship_cabins$FrontNeighborsTransportedPct[i] <- ifelse(nrow(front_neighbor) == 0, 0, front_neighbor$CabinTransportedPct)
    ship_cabins$BackNeighbors[i] <- ifelse(nrow(back_neighbor) == 0, 0, back_neighbor$CabinNumTransported + back_neighbor$CabinNumNotTransported)
    ship_cabins$BackNeighborsTransported[i] <- ifelse(nrow(back_neighbor) == 0, 0, back_neighbor$CabinNumTransported)
    ship_cabins$BackNeighborsTransportedPct[i] <- ifelse(nrow(back_neighbor) == 0, 0, back_neighbor$CabinTransportedPct)
    ship_cabins$DiagFrontNeighbors[i] <- ifelse(nrow(diag_front_neighbor) == 0, 0, diag_front_neighbor$CabinNumTransported + diag_front_neighbor$CabinNumNotTransported)
    ship_cabins$DiagFrontNeighborsTransported[i] <- ifelse(nrow(diag_front_neighbor) == 0, 0, diag_front_neighbor$CabinNumTransported)
    ship_cabins$DiagFrontNeighborsTransportedPct[i] <- ifelse(nrow(diag_front_neighbor) == 0, 0, diag_front_neighbor$CabinTransportedPct)
    ship_cabins$DiagBackNeighbors[i] <- ifelse(nrow(diag_back_neighbor) == 0, 0, diag_back_neighbor$CabinNumTransported + diag_back_neighbor$CabinNumNotTransported)
    ship_cabins$DiagBackNeighborsTransported[i] <- ifelse(nrow(diag_back_neighbor) == 0, 0, diag_back_neighbor$CabinNumTransported)
    ship_cabins$DiagBackNeighborsTransportedPct[i] <- ifelse(nrow(diag_back_neighbor) == 0, 0, diag_back_neighbor$CabinTransportedPct)
}
rm(splits, front_neighbor, side_neighbor, back_neighbor, diag_front_neighbor, diag_back_neighbor)
```

### Add Cabin Columns
```{r}
ship <- merge(ship, ship_cabins, by=c('Cabin', 'Side', 'Deck', 'Num'), all.x=T)
```

### GID is independent of test/train split
```{r}
bin_width <- nrow(ship_groups) / 20
d <- rorschach(null_permute('GroupTrain'), ship_groups, n = 20, p = 0)
ggplot(data=d[!d$GroupTrain, ], aes(x=GID)) +
  geom_histogram(breaks=seq(min(ship_groups$GID) + bin_width, by = bin_width, length.out = 19)) +
  facet_wrap(~ .sample)

d <- lineup(null_permute('GroupTrain'), ship_groups, n = 20)
ggplot(data=d[!d$GroupTrain, ], aes(x=GID)) +
  geom_histogram(breaks=seq(min(ship_groups$GID) + bin_width, by = bin_width, length.out = 19)) +
  facet_wrap(~ .sample)

attr(d, "pos")

pvisual(0, 1)
```

### Percentage of Cabins Having the Same CryoSleep

```{r}
count <- 0
total <- 0
for (i in unique(ship$Cabin[!is.na(ship$Cabin)])){
  g <- ship[!is.na(ship$CryoSleep) & !is.na(ship$Cabin) & ship$Cabin == i, ]
  if (nrow(g) <= 1) next
  if (length(unique(g$CryoSleep)) == 1) {
    count <- count + 1
    total <- total + 1
  }
  else if (length(unique(g$CryoSleep)) == 2) {
    total <- total + 1
  }
}
print(paste0(count, " / ", total, " = ", round(count/total, 2)))
rm(g)
```

### Groups Share the Same HomePlanet (But Not Destination)
#### Evidence
```{r}
table(ship_groups$GroupNumHomePlanets)
table(ship_groups$GroupNumDestinations)
```

#### Update
```{r}
for (i in 1:nrow(ship)){
    if (is.na(ship$HomePlanet[i]) && !is.na(ship$GroupHomePlanetMode[i])){
        ship$HomePlanet[i] <- ship$GroupHomePlanetMode[i]
    }
}
```

# Write ship to CSV
```{r}
write.csv(ship, 'data/ship.csv', row.names=F)
```
